2022/11/16
Added information for Himari.
Added the ability to set student buffs on a student's stat display screen.
A student's own buffs as well as buffs that can be applied from other students can be set.
For skills that grant stacking or conditional buffs, the stack count/condition state can be toggled by clicking on the stack counter.
When selecting two skills that grant buffs that overwrite each other, a warning will be displayed and only the first effect will be applied.